Change screens routes prefix during rename phase + cleaning files according to our code style
