+++ src/config.ts
@@ -1,5 +1,6 @@
 import { NativeModules } from 'react-native';
 import config from 'react-native-config';
+
 const env = require('./config.env.js').default;
 
 // Combine native config and generated JS config
