+++ src/screens/counter/Counter.tsx
@@ -1,14 +1,17 @@
 import * as React from 'react';
 import { View, Text } from 'react-native';
 import { observer } from 'mobx-react';
-import Button from 'components/button/Button';
-import { UI } from 'store';
-import * as styles from './Counter.css';
 
-interface IProps {
-  componentId: string;
+import { UI } from 'store';
+import { IReactNavigation } from 'typings';
+
+import Button from 'components/button/Button';
+
+interface IProps extends IReactNavigation {
   testID?: string;
 }
+
+const s = require('./Counter.css');
 
 @observer
 export default class Counter extends React.Component<IProps> {
@@ -29,23 +32,26 @@ export default class Counter extends React.Component<IProps> {
 
   render() {
     return (
-      <View style={styles.host} testID="COUNTER_SCREEN">
-        <View style={styles.content}>
-          <Text style={styles.text}>Counter: {UI.counter}</Text>
+      <View style={s.host} testID="COUNTER_SCREEN">
+        <View style={s.content}>
+          <Text style={s.text}>Counter: {UI.counter}</Text>
         </View>
-        <View style={styles.actions}>
+
+        <View style={s.actions}>
           <Button
             title="Decrement"
             onPress={UI.decrement}
             testID="BUTTON_DECREMENT"
-            style={styles.actions__button}
+            style={s.actions__button}
           />
-          <View style={styles.actions__spacer} />
+
+          <View style={s.actions__spacer} />
+
           <Button
             title="Increment"
             onPress={UI.increment}
             testID="BUTTON_INCREMENT"
-            style={styles.actions__button}
+            style={s.actions__button}
           />
         </View>
       </View>
