+++ src/screens/home/Home.tsx
@@ -1,16 +1,19 @@
 import * as React from 'react';
 import { View, Text } from 'react-native';
 import { observer } from 'mobx-react';
-import Button from 'components/button/Button';
-import { UI } from 'store';
 import { Navigation } from 'react-native-navigation';
-import { COUNTER } from 'screens';
-import * as styles from './Home.css';
 
-interface IProps {
-  componentId: string;
+import { UI } from 'store';
+import { COUNTER } from 'screens';
+import { IReactNavigation } from 'typings';
+
+import Button from 'components/button';
+
+interface IProps extends IReactNavigation {
   testID?: string;
 }
+
+const s = require('./Home.css');
 
 @observer
 export default class Home extends React.Component<IProps> {
@@ -39,10 +42,11 @@ export default class Home extends React.Component<IProps> {
 
   render() {
     return (
-      <View style={styles.host} testID="HOME_SCREEN">
-        <View style={styles.content}>
-          <Text style={styles.text}>Welcome Home</Text>
+      <View style={s.host} testID="HOME_SCREEN">
+        <View style={s.content}>
+          <Text style={s.text}>Welcome Home</Text>
         </View>
+
         <Button onPress={this.onCounterScreenPress} title="Counter Screen" />
       </View>
     );
