+++ src/screens/index.ts
@@ -1,11 +1,13 @@
 import { Navigation } from 'react-native-navigation';
-import Home from './home/Home';
-import Counter from './counter/Counter';
+
+import Home from './home';
+import Counter from './counter';
 
 export const HOME = 'ueno-rns.Home';
 export const COUNTER = 'ueno-rns.Counter';
 
 export const Screens = new Map();
+
 Screens.set(HOME, Home);
 Screens.set(COUNTER, Counter);
 
@@ -15,9 +17,7 @@ export const startApp = () => {
       stack: {
         id: 'ROOT_STACK',
         children: [{
-          component: {
-            name: HOME,
-          },
+          component: { name: HOME },
         }],
       },
     },
