+++ src/typings/index.ts
@@ -1,0 +1,7 @@
+interface IReactNavigation {
+  componentId: string;
+}
+
+export {
+  IReactNavigation,
+};
